Update README.md Python Calculator
In this program, the user is prompted to select an operation, with options 1, 2, 3, and 4 considered valid. If any other input is provided, the program displays "Invalid Input" and continues looping until a valid option is chosen. Then, the program prompts the user to enter two numbers. Using an if-elif-else branching structure, the program executes the corresponding section based on the chosen operation. There are user-defined functions add(), subtract(), multiply(), and divide() to evaluate the respective operations and display the output.